+++ CJBCheatsMenu/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * 'Always auto-feed' now works without silos if 'infinite hay' is enabled.
+* Internal refactor to translation handling, and now uses game translations where possible.
 
 ## 1.25.2
 Released 12 September 2020 for SMAPI 3.7 or later.